TSC Command not found error, Exception gets thrown by a node_module?
This is just a second error text to bugfix and for us to both look at. I don't know why it doesn't work. I also now realize that this will automatically generate TLS certificate for us if we don't have it. Just take a look at it for me because I am really confused as to where to start when bugfixing this.